move schema to job/schema.py
